requirements.txt
blinker==1.9.0
click==8.1.8
et_xmlfile==2.0.0
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
openpyxl==3.1.5
pathlib==1.0.1
pyee==11.1.1
python-dotenv==1.0.1
slack_sdk==3.34.0
slackeventsapi==3.0.3
typing_extensions==4.12.2
Werkzeug==3.1.3
